ChemicalElement: changed the type of yearDiscovered back to string because some years are not specified, e.g., 'ancient' is the year for carbon. PeriodicTable: changed accordingly.

+++ src/main/kotlin/graphics/scenery/PeriodicTable.kt
@@ -78,7 +78,7 @@ open class PeriodicTable {
             boilingPoint = pureStringElement.cell[13].toFloatOrNull(),
             density = pureStringElement.cell[14].toFloatOrNull(),
             groupBlock = pureStringElement.cell[15],
-            yearDiscovered = pureStringElement.cell[16].toInt()
+            yearDiscovered = pureStringElement.cell[16]
             )
             elementList.add(element)
         }
